changed response to messages and refactored order of methods
